requirements.txt
telethon==1.37.0
python-dotenv==1.0.1